% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACBDStats.R
\docType{class}
\name{ACBDStats-class}
\alias{ACBDStats-class}
\alias{ACBDStats}
\title{Concrete interface for ACBD designs}
\description{
Specialized Concrete Definition.
}
\section{Fields}{

\describe{
\item{\code{report}}{list or analytical summary.}
}}


\examples{
data(wheat, package = "yabaf")
b <- Breeder(wheat)
a <- ACBDStats(b)
a$report$h2
}
\seealso{
[Breeder() and SplitPlot()] for the pure virtual and concrete
    abstract template.
}
