% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Augmented.R
\docType{class}
\name{Augmented-class}
\alias{Augmented-class}
\alias{Augmented}
\title{Concrete interface for Augmented designs}
\description{
Specialized Concrete Definition.
}
\section{Fields}{

\describe{
\item{\code{report}}{list or analytical summary.}
}}


\examples{
data("burgueno.unreplicated", package = "agridat")
dat <- burgueno.unreplicated
names(dat)[names(dat) == "gen"] <- "treat"
names(dat)[names(dat) == "col"] <- "column"
names(dat)[names(dat) == "yield"] <- "response"
dat$role <- ifelse(dat$treat == "G000", "check", "test")
b <- Breeder(dat)
a <- Augmented(b)
a$report$fit$sigma
}
\seealso{
[Breeder() and SplitPlot()] for the pure virtual and concrete
    abstract template.
}
